requirements.txt
# Core Python Libraries
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0

# Machine Learning and AI
scikit-learn>=1.1.0
xgboost>=1.6.0
langchain>=0.1.0
openai>=1.0.0

# Web and API Integration
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# Advanced AI Models
groq>=0.4.0
crewai>=0.22.0

# Web Framework
streamlit>=1.10.0
fastapi>=0.78.0
uvicorn>=0.17.6

# Visualization
plotly>=5.8.0
matplotlib>=3.5.0

# Data Processing
pyarrow>=8.0.0
pydantic>=1.9.0

# Monitoring and Logging
python-dotenv>=0.20.0
sentry-sdk>=1.5.0
structlog>=23.1.0
logging>=0.5.1.2

# Computer Vision and OCR
opencv-python>=4.6.0
pytesseract>=0.3.9
pillow>=9.1.0

# Async Programming
asyncio>=3.4.3

# Optional: Advanced Features
spacy>=3.3.0
nltk>=3.7

# Development and Testing
pytest>=7.1.0

# Notes:
# 1. Some libraries may require specific installation methods
# 2. Versions are minimum requirements
# 3. Consider using virtual environments
